mobile/android/tests/browser/chrome/test_get_last_visited.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1214366
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1214366</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript;version=1.7">

  "use strict";

  const { classes: Cc, interfaces: Ci, utils: Cu } = Components;

  Cu.import("resource://gre/modules/Services.jsm");
  Cu.import("resource://gre/modules/Messaging.jsm");
  Cu.import("resource://gre/modules/Task.jsm");

  let chromeWin = Services.wm.getMostRecentWindow("navigator:browser");
  let BrowserApp = chromeWin.BrowserApp;

  function get_last_visited(prePath) {
    return Messaging.sendRequestForResult({
      type: "History:GetPrePathLastVisitedTimeMilliseconds",
      prePath: prePath,
    });
  };

  var browser = BrowserApp.addTab("about:blank").browser;

  SimpleTest.registerCleanupFunction(function cleanup() {
    BrowserApp.closeTab(BrowserApp.getTabForBrowser(browser));
  });

  // N.b.: the write to the Fennec DB happens before the Gecko notification
  // is fired.  This is delicate.
  function add_history_visit(url) {
    browser.loadURI(url, null, null);
    return promiseLinkVisit(url);
  };

  // Be aware that mochi.test:8888 redirects, while example.org does
  // not.  The redirects can impact this test, since they can write to
  // the history database.
  const url1 = "http://example.org/chrome/mobile/android/tests/browser/chrome/basic_article.html";
  const url2 = "http://example.org/chrome/mobile/android/tests/browser/chrome/video_discovery.html";

  add_task(function* test_get_last_visited() {
    var v = yield get_last_visited("https://random.com/");
    is(v, 0, "Last visited timestamp is 0 for unknown prePath.");

    let prePath = Services.io.newURI(url1, null, null).prePath + "/";
    is(prePath, Services.io.newURI(url2, null, null).prePath + "/", "url1 and url2 have the same prePath");

    let t0 = Date.now();
    yield add_history_visit(url1);
    let t1 = Date.now();
    v = yield get_last_visited(prePath);
    ok(t0 <= v, "Last visited timestamp is after visit.");
    ok(v <= t1, "Last visited timestamp is before present");

    let t2 = Date.now();
    yield add_history_visit(url1);
    v = yield get_last_visited(prePath);
    ok(t2 <= v, "Last visited timestamp is updated after visit.");

    let t3 = Date.now();
    yield add_history_visit(url2);
    v = yield get_last_visited(prePath);
    ok(t3 <= v, "Last visited timestamp is updated after visit to URL with same prePath.");
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1214366">Mozilla Bug 1214366</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
